added negative tests
